Apply comments from code review
